+++ include/constant/constants.h
@@ -3,6 +3,7 @@
 namespace constant {
 float kMaxVelocity = 50.0f;  // in/s
 
+
 }  // namespace constant
 
 /*#ifndef _CONSTANTS_H_
